requirements.txt
streamlit>=1.28.0
google-adk>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
google-generativeai>=0.8.0
groq>=0.4.1
